MLX:47 pyswitch API for ping utility
